% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{run Shiny Example}
\usage{
runShiny(shinyappname = "MassNBCtools")
}
\arguments{
\item{shinyappname}{Shiny appplication name, default = MassNBCtools}
}
\description{
Launches Shiny app for MassNBCtools package.
}
\details{
The Shiny app based on the R package MassNBCtools is included in
the R package. This function launches that app.

The Shiny app is online at:
https://tetratech-wtr-wne.shinyapps.io/MassNBCtools
}
\examples{
\dontrun{
# Run Function (full EMVL version, default)
runShiny()

}
}
