% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{run Shiny Example}
\usage{
runShiny(shinyappname = "MassDEP_NBCtools")
}
\arguments{
\item{shinyappname}{Shiny appplication name, default = MassDEP_NBCtools}
}
\description{
Launches Shiny app for MassDEP_NBCtools package.
}
\details{
The Shiny app based on the R package MassDEP_NBCtools is included in
the R package. This function launches that app.

The Shiny app is online at:
https://tetratech-wtr-wne.shinyapps.io/MassDEP_NBCtools
}
\examples{
\dontrun{
# Run Function (full EMVL version, default)
runShiny()

}
}
